Create setup_common.py
